+++ CHANGELOG.md
@@ -5,6 +5,8 @@ All notable changes to this project will be documented in this file. This projec
 
 ## Unreleased
 
+## [5.2.0] - 2026-01-05
+
 ### Added
 
 - New domain event dispatcher features:
@@ -614,6 +616,10 @@ All notable changes to this project will be documented in this file. This projec
 ## [0.1.0] - 2023-11-18
 
 Initial release.
+
+[5.2.0]: https://github.com/cloudcreativity/ddd-modules/compare/v5.1.0...v5.2.0
+
+[5.1.0]: https://github.com/cloudcreativity/ddd-modules/compare/v5.0.0...v5.1.0
 
 [5.0.0]: https://github.com/cloudcreativity/ddd-modules/compare/v5.0.0-rc.4...v5.0.0